Refactor in preparation of additional testing
